Utilities will ignore the not-shelved files for dark archive

+++ lib/cdx_merge_sort_publish_service.rb
@@ -34,8 +34,8 @@ module Dor
       
       def clean
         FileUtils.mv(@source_cdx_dir, @cdx_backup_directory)
-        FileUtils.rm(@working_sorted_duplicate_cdx)
-        FileUtils.rm(@working_merged_cdx)
+        FileUtils.rm(@working_sorted_duplicate_cdx) if File.exist?(@working_sorted_duplicate_cdx)
+        FileUtils.rm(@working_merged_cdx) if File.exist?(@working_merged_cdx)
       end
     end
   end
